Struct nphysics2df32​::detection​::joint​::BallInSocket​

pub struct BallInSocket {
    // some fields omitted
}

A ball-in-socket joint.

This is usually used to create ragdolls.

Methods

impl BallInSocket

fn new(anchor1: Anchor<Vect>, anchor2: Anchor<Vect>) -> BallInSocket

Creates a ball-in-socket joint.

fn up_to_date(&self) -> bool

Tells if this joint has been modified by the user.

fn set_local1(&mut self, local1: Vect)

Sets the the second anchor position.

The position is expressed in the second attached body’s local coordinates.

fn set_local2(&mut self, local2: Vect)

Sets the the second anchor position.

The position is expressed in the second attached body’s local coordinates.