Struct nphysics2df32​::detection​::joint​::JointManager​

pub struct JointManager {
    // some fields omitted
}

Structure that handles creation and removal of joints.

Methods

impl JointManager

fn new() -> JointManager

Creates a new JointManager.

fn joints<'a>(&'a self) -> &'a HashMap<uint, Constraint, UintTWHash>

Joints handled by this manager.

fn joints_with_body<'a>(&'a self, body: &Rc<RefCell<RigidBody>>) -> Option<&'a [Constraint]>

List of joints attached to a specific body.

fn add_ball_in_socket(&mut self, joint: Rc<RefCell<BallInSocket>>, activation: &mut ActivationManager)

Add a BallInSocket joint to this manager.

This will force the activation of the two objects attached to the joint.

fn remove_ball_in_socket(&mut self, joint: &Rc<RefCell<BallInSocket>>, activation: &mut ActivationManager)

Removes a BallInSocket joint from this manager.

This will force the activation of the two objects attached to the joint.

fn add_fixed(&mut self, joint: Rc<RefCell<Fixed>>, activation: &mut ActivationManager)

Add a Fixed joint to this manager.

This will force the activation of the two objects attached to the joint.

fn remove_joint<T: Joint<M>, M>(&mut self, joint: &Rc<RefCell<T>>, activation: &mut ActivationManager)

Removes a joint from this manager.

This will force the activation of the two objects attached to the joint.

fn remove(&mut self, b: &Rc<RefCell<RigidBody>>, activation: &mut ActivationManager)

Removes every joint attached to a given rigid body.

This will force the activation of every object attached to the deleted joints.