Struct nphysics2df32​::detection​::BodyBodyDispatcher​

pub struct BodyBodyDispatcher {
    // some fields omitted
}

Collision detector dispatcher for rigid bodies.

This is meat to be used as the broad phase collision dispatcher.

Methods

impl BodyBodyDispatcher

fn new(d: Rc<GeomGeomDispatcher>) -> BodyBodyDispatcher

Creates a new BodyBodyDispatcher given a dispatcher for pairs of rigid bodies' geometry.