Trait nphysics::integration::integrator::Integrator

pub trait Integrator<O> {
    fn update(&mut self, dt: Scalar, o: &mut O);
}

Trait implemented by every integrator.

An integrator is a structure capable of updating a dynamic body position and orientation after a given time-step.

Required Methods

fn update(&mut self, dt: Scalar, o: &mut O)

Updates the position and orientation of the object o after a time step of dt.