Struct nphysics2df32​::integration​::BodyDamping​

pub struct BodyDamping {
    // some fields omitted
}

A linear and angular velocity damper.

This will remove a part of the linear and angular velocity of every rigid body at each frame. Do not use unless you are experiencing unrealistic vibrations or very unstable joints.

Methods

impl BodyDamping

fn new(linear_damping: Scalar, angular_damping: Scalar) -> BodyDamping

Creates a new BodyDamping.

Arguments:

linear_damping - coefficient in [0, 1] the linear velocity of each rigid body is multiplied at each update.

linear_damping - coefficient in [0, 1] the angular velocity of each rigid body is multiplied at each update.