Struct nphysics2df32​::object​::RigidBody​

pub struct RigidBody {
    // some fields omitted
}

The rigid body structure.

This is the structure describing an object on the physics world.

Methods

impl RigidBody

fn transform_ref<'r>(&'r self) -> &'r Matrix

Gets a reference to this body's transform.

fn geom_ref<'r>(&'r self) -> &'r Geom

Gets a reference to this body's geometry.

fn geom<'r>(&'r self) -> Arc<Box<Geom + Send + Sync>>

Gets a copy of this body's shared geometry.

fn center_of_mass<'r>(&'r self) -> &'r Vect

Gets a reference to this body's center of mass.

fn restitution(&self) -> Scalar

Gets this body's restitution coefficent.

The actual restitution coefficient of a contact is computed averaging the two bodies friction coefficient.

fn friction(&self) -> Scalar

Gets this body's friction coefficient.

The actual friction coefficient of a contact is computed averaging the two bodies friction coefficient.

fn is_active(&self) -> bool

Indicates whether or not this rigid body is active.

An inactive rigid body is a body that did not move for some time. It is not longer simulated by the physics engine as long as no other object touches it. It is automatically activated by the physics engine.

fn deactivation_threshold(&self) -> Option<Scalar>

The velocity threshold bellow whith the rigid body might be deactivated.

If None, the object cannot be deactivated. If the total squared velocity (i-e: v2 + w2) falls bellow this threshold for a long enough time, the rigid body will fall asleep (i-e be "forzen") for performance reasons.

fn set_deactivation_threshold(&mut self, threshold: Option<Scalar>)

Set the velocity threshold bellow whith the rigid body might be deactivated.

If None, the object cannot be deactivated. If the total squared velocity (i-e: v2 + w2) falls bellow this threshold for a long enough time, the rigid body will fall asleep (i-e be "forzen") for performance reasons.

fn new_dynamic<G: Send + Sync + Geom + Volumetric>(geom: G, density: Scalar, restitution: Scalar, friction: Scalar) -> RigidBody

Creates a new rigid body that can move.

fn new_static<G: Send + Sync + Geom>(geom: G, restitution: Scalar, friction: Scalar) -> RigidBody

Creates a new rigid body that cannot move.

fn new(geom: Arc<Box<Geom + Send + Sync>>, mass_properties: Option<(Scalar, Vect, AngularInertia)>, restitution: Scalar, friction: Scalar) -> RigidBody

Creates a new rigid body with a given geometry.

Use this if the geometry is shared by multiple rigid bodies. Set mass_properties to None if the rigid body is to be static.

fn can_move(&self) -> bool

Indicates whether this rigid body is static or dynamic.

fn lin_acc_scale(&self) -> Vect

Gets the linear acceleraction scale of this rigid body.

fn set_lin_acc_scale(&mut self, scale: Vect)

Sets the linear acceleration scale of this rigid body.

fn ang_acc_scale(&self) -> Orientation

Gets the angular acceleration scale of this rigid body.

fn set_ang_acc_scale(&mut self, scale: Orientation)

Sets the angular acceleration scale of this rigid body.

fn lin_vel(&self) -> Vect

Get the linear velocity of this rigid body.

fn set_lin_vel(&mut self, lv: Vect)

Sets the linear velocity of this rigid body.

fn lin_acc(&self) -> Vect

Gets the linear acceleration of this rigid body.

fn set_lin_acc(&mut self, lf: Vect)

Sets the linear acceleration of this rigid body.

Note that this might be reset by the physics engine automatically.

fn ang_vel(&self) -> Orientation

Gets the angular velocity of this rigid body.

fn set_ang_vel(&mut self, av: Orientation)

Sets the angular velocity of this rigid body.

fn ang_acc(&self) -> Orientation

Gets the angular acceleration of this rigid body.

fn set_ang_acc(&mut self, af: Orientation)

Sets the angular acceleration of this rigid body.

Note that this might be reset by the physics engine automatically.

fn inv_mass(&self) -> Scalar

Gets the inverse mass of this rigid body.

fn set_inv_mass(&mut self, m: Scalar)

Sets the inverse mass of this rigid body.

fn inv_inertia<'r>(&'r self) -> &'r AngularInertia

Gets the inverse inertia tensor of this rigid body.

fn set_inv_inertia(&mut self, ii: AngularInertia)

Sets the inverse inertia tensor of this rigid body.

Not that this is reset at every update by the physics engine.