Module nphysics::volumetric

Volume and inertia tensor computation.

Traits

InertiaTensor
Trait implemented by inertia tensors.

Volumetric
Trait implemented by objects which have a mass, a center of mass, and an inertia tensor.

Functions

ball_center_of_mass
The center of mass of a ball.

ball_surface
The surface of a ball.

ball_unit_angular_inertia
The unit angular inertia of a ball.

ball_volume
The volume of a ball.

cone_center_of_mass
The center of mass of a cone.

cone_surface
The surface of a cone.

cone_unit_angular_inertia
The unit angular inertia of a cone.

cone_volume
The volume of a cone.

convex_hull_center_of_mass
The center of mass of the convex hull of a set of points.

convex_hull_surface
The surface of a convex hull.

convex_hull_unit_angular_inertia
The angular inertia of the convex hull of a set of points.

convex_hull_volume
The volume of the convex hull of a set of points.

convex_mesh_mass_properties
The mass properties of a convex mesh.

convex_mesh_surface
The surface of a convex mesh.

convex_mesh_volume_and_center_of_mass
The volume and center of mass of a convex mesh.

cuboid_center_of_mass
The center of mass of a cuboid.

cuboid_surface
The surface of a cuboid.

cuboid_unit_angular_inertia
The unit angular inertia of a cuboid.

cuboid_volume
The volume of a cuboid.

cylinder_center_of_mass
The center of mass of a cylinder.

cylinder_surface
The surface of a cylinder.

cylinder_unit_angular_inertia
The unit angular inertia of a cylinder.

cylinder_volume
The volume of a cylinder.