Struct nphysics::integration::BodyForceGenerator

pub struct BodyForceGenerator {
    // some fields omitted
}

A constant linear and angular force generator.

Methods

impl BodyForceGenerator

fn new(lin_acc: Vect, ang_acc: Orientation) -> BodyForceGenerator

Creates a new BodyForceGenerator.

Arguments:

lin_acc - the linear acceleration to apply to every body on the scene.

ang_acc - the angular acceleration to apply to every body on the scene.

impl BodyForceGenerator

fn lin_acc(&self) -> Vect

The linear acceleration applied by this force generator.

fn set_lin_acc(&mut self, lin_acc: Vect)

Sets the linear acceleration applied by this force generator.

fn ang_acc(&self) -> Orientation

The angular acceleration applied by this force generator.

fn set_ang_acc(&mut self, ang_acc: Orientation)

Sets the angular acceleration applied by this force generator.